% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sensors_metadata.R
\name{build_sensors_metadata}
\alias{build_sensors_metadata}
\title{Get metadata for sensors}
\usage{
build_sensors_metadata()
}
\description{
metadata is going to be used to make queries against the api. \cr
This function retrieves SCODE and TYPE for each sensor and thus indicates from \cr
what sensor we want to retrieve data and what datapoint we want to query.
}
