% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurement_data.R
\name{get_measurement_data}
\alias{get_measurement_data}
\title{Function for retrieving data from weather API}
\usage{
get_measurement_data(start_date, end_date, scode, type)
}
\arguments{
\item{start_date}{a start_date}

\item{end_date}{an end_date}

\item{scode}{a sensor code see, sensors dataset}

\item{type}{a measurement specification, see sensors dataset}
}
\description{
Function for retrieving data from weather API
}
