% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_measurement.R
\docType{data}
\name{sensor_measurement}
\alias{sensor_measurement}
\title{Metadata: description of TYPE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.
}
\usage{
sensor_measurement
}
\description{
Metadata: description of TYPE
}
\keyword{datasets}
