% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\docType{data}
\name{sensors}
\alias{sensors}
\title{Sensors dataset
This dataset contains all necessary metadata to make queries on the South Tyrolian weather API.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 842 rows and 15 columns.
}
\usage{
sensors
}
\description{
Sensors dataset
This dataset contains all necessary metadata to make queries on the South Tyrolian weather API.
}
\keyword{datasets}
