% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_codes.R
\docType{data}
\name{station_codes}
\alias{station_codes}
\title{Description of station}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 167 rows and 2 columns.
}
\usage{
station_codes
}
\description{
Description of station
}
\keyword{datasets}
